Add separate alert.ts and refactor close button
