+++ index.html
@@ -73,7 +73,9 @@
     </header>
 
     <div class="alert" id="no-requirements-alert">
-      <span class="closebtn" onclick="this.parentElement.style.display='none';"
+      <span
+        class="alert-close-button"
+        onclick="this.parentElement.style.display='none';"
         >&times;</span
       >
       Cities displayed may still have parking requirements for a few specific
